requirements.txt
# Core Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
djangorestframework-simplejwt[crypto]>=5.3.0

# Database
psycopg2-binary>=2.9.0

# Environment
python-dotenv>=1.0.0

# Task scheduling (for bonus email features)
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0

# CORS (if needed for frontend)
django-cors-headers>=4.3.0

# API Documentation
drf-spectacular>=0.27.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=20.0.0

# Code Quality
black>=23.12.0
flake8>=7.0.0
isort>=5.13.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0

# Security
cryptography>=41.0.0
